Split application request routing
