+++ scripts/mutation/equivalent-mutants.txt
@@ -930,7 +930,7 @@ src/shared/dates.ts:432:30  1 → 0    # bookedRangeLabel's `fallbackDurationDay
 # Admin segment dispatch (admin/index.ts) — two provably-equivalent survivors
 # from the routeAdmin/adminPathSegment run over the manifest + footer suites.
 src/features/admin/index.ts:257:21  ?? → ||   # `path.split("/")[2] ?? ""`: the RHS is "" and the only falsy string is "", so `x ?? ""` and `x || ""` agree on every string|undefined input
-src/features/admin/index.ts:299:55  return null → return undefined   # routeAdmin's only consumer folds the result with `?? notFoundResponse()` (features/index.ts routeMainApp), and ?? treats null and undefined identically
+src/features/admin/index.ts:299:55  return null → return undefined   # routeAdmin's only consumer folds the result with `?? notFoundResponse()` (features/app/routes.ts routeMainApp), and ?? treats null and undefined identically
 
 # Logger (logger.ts) — two provably-equivalent survivors from the run over the
 # logger suites.
