+++ src/features/app/request.ts
@@ -1,0 +1,270 @@
+/* jscpd:ignore-start -- imports */
+import { once } from "#fp";
+import { SessionKeyError } from "#routes/auth.ts";
+import {
+  applySecurityHeaders,
+  contentTypeRejectionResponse,
+  isEmbeddablePath,
+  isValidContentType,
+} from "#routes/middleware.ts";
+import { bufferRequestBody } from "#routes/request-body.ts";
+import {
+  databaseBusyResponse,
+  migrationInProgressResponse,
+  redirectResponse,
+  siteNotActivatedResponse,
+  temporaryErrorResponse,
+  withCookie,
+} from "#routes/response.ts";
+import { getPrefix, settingsForPath } from "#routes/settings-bundles.ts";
+import { routeStatic } from "#routes/static.ts";
+import type { ServerContext } from "#routes/types.ts";
+import { parseCookies, parseRequest } from "#routes/url.ts";
+import {
+  loadEffectiveDomain,
+  seedEffectiveDomainHost,
+} from "#shared/config.ts";
+import {
+  clearFlashCookie,
+  clearSessionCookie,
+  parseFlashValue,
+} from "#shared/cookies.ts";
+import { maybeBackfillActivityLog } from "#shared/db/activity-log-backfill.ts";
+import { DatabaseBusyError } from "#shared/db/client.ts";
+import {
+  initDb,
+  MigrationInProgressError,
+  MissingSettingsTableError,
+} from "#shared/db/migrations.ts";
+import { maybeRunPrunes } from "#shared/db/prune.ts";
+import { enableQueryLog } from "#shared/db/query-log.ts";
+import { settings } from "#shared/db/settings.ts";
+import { assertSettingsReadsDeclared } from "#shared/db/settings-audit.ts";
+import { hasFlash, setFlashContext } from "#shared/flash-context.ts";
+import { FormParams } from "#shared/form-data.ts";
+import { takeForm } from "#shared/form-stash.ts";
+import { clearSavedFormData, setSavedFormData } from "#shared/forms.tsx";
+import { detectIframeMode } from "#shared/iframe.ts";
+import {
+  createRequestTimer,
+  ErrorCode,
+  formatRequestError,
+  logError,
+  logRequest,
+} from "#shared/logger.ts";
+import { addPendingWork, flushPendingWork } from "#shared/pending-work.ts";
+import { getRethrowErrors } from "#shared/test-overrides.ts";
+import { defineAppRoute, routeMainApp } from "./routes.ts";
+import {
+  ensureCustomCssResponse,
+  isSetupPath,
+  shouldBufferRequestBody,
+  trackingRedirectLocation,
+} from "./rules.ts";
+
+/* jscpd:ignore-end */
+
+const loadSetupRoutes = once(async () =>
+  (await import("#routes/setup.ts")).createSetupRouter(
+    settings.setup.isComplete,
+  ),
+);
+
+const handleRequestInternal = defineAppRoute(
+  async ({ request, path, method, server }) => {
+    if (isSetupPath(path)) {
+      const setupResponse = await (await loadSetupRoutes())(
+        request,
+        path,
+        method,
+      );
+      if (setupResponse) return setupResponse;
+    }
+
+    if (!(await settings.setup.isComplete())) {
+      return isSetupPath(path)
+        ? redirectResponse("/setup")
+        : siteNotActivatedResponse();
+    }
+
+    return routeMainApp({ method, path, request, server });
+  },
+);
+
+const initializeDatabaseForPath = async (
+  path: string,
+): Promise<Response | null> => {
+  try {
+    await initDb({ allowMissingSettings: isSetupPath(path) });
+    return null;
+  } catch (error) {
+    if (error instanceof MissingSettingsTableError) {
+      return siteNotActivatedResponse();
+    }
+    if (error instanceof MigrationInProgressError) {
+      return migrationInProgressResponse();
+    }
+    throw error;
+  }
+};
+
+const logAndReturn = (
+  response: Response,
+  method: string,
+  path: string,
+  getElapsed: () => number,
+): Response => {
+  logRequest({
+    durationMs: getElapsed(),
+    method,
+    path,
+    status: response.status,
+  });
+  return response;
+};
+
+const applyFlashFromCookie = (request: Request, url: URL): string | null => {
+  const flashId = url.searchParams.get("flash");
+  const flashRaw = flashId
+    ? parseCookies(request).get(`flash_${flashId}`)
+    : null;
+  const flash = flashRaw ? parseFlashValue(flashRaw) : null;
+  if (flash) setFlashContext(flash);
+
+  if (flash?.formToken) {
+    const stashed = takeForm(flash.formToken);
+    if (stashed) setSavedFormData(new FormParams(stashed));
+  }
+  return flash ? flashId : null;
+};
+
+const prepareRequestEnvironment = async (
+  url: URL,
+  path: string,
+  method: string,
+): Promise<void> => {
+  if (method === "GET" && getPrefix(path) === "admin") enableQueryLog();
+
+  await settings.loadKeys(settingsForPath(path));
+
+  if (!(method === "POST" && path === "/admin/privacy/orphans")) {
+    addPendingWork(maybeRunPrunes());
+  }
+  addPendingWork(maybeBackfillActivityLog());
+  loadEffectiveDomain(url);
+};
+
+const routeAndFinalize = async (
+  request: Request,
+  url: URL,
+  path: string,
+  method: string,
+  server: ServerContext | undefined,
+): Promise<Response> => {
+  const embeddable = isEmbeddablePath(path);
+  const consumedFlashId = applyFlashFromCookie(request, url);
+  const response = await handleRequestInternal(request, path, method, server);
+
+  if (consumedFlashId && hasFlash()) {
+    withCookie(response, clearFlashCookie(consumedFlashId));
+  }
+  return applySecurityHeaders(response, embeddable);
+};
+
+const handleRoutingError = (
+  error: unknown,
+  method: string,
+  path: string,
+): Response => {
+  if (error instanceof DatabaseBusyError) {
+    logError({
+      code: ErrorCode.DB_BUSY,
+      detail: formatRequestError(method, path, error),
+      error,
+    });
+    return databaseBusyResponse(["GET", "HEAD"].includes(method));
+  }
+
+  logError({
+    code: ErrorCode.CDN_REQUEST,
+    detail: formatRequestError(method, path, error),
+    error,
+  });
+  if (
+    getRethrowErrors() &&
+    !(error instanceof SessionKeyError) &&
+    !Deno.env.get("TEST_EXPECT_ERROR")
+  ) {
+    throw error;
+  }
+  if (error instanceof SessionKeyError) {
+    return redirectResponse("/admin", clearSessionCookie());
+  }
+  return temporaryErrorResponse();
+};
+
+/** Run the request pipeline inside the request-scoped contexts from index.ts. */
+export const processRequest = async (
+  request: Request,
+  server: ServerContext | undefined,
+): Promise<Response> => {
+  const { url, path, method } = parseRequest(request);
+  const getElapsed = createRequestTimer();
+  detectIframeMode(url);
+  clearSavedFormData();
+
+  const finish = (response: Response): Response =>
+    logAndReturn(
+      ensureCustomCssResponse(path, response),
+      method,
+      path,
+      getElapsed,
+    );
+
+  let response!: Response;
+  try {
+    const bufferedRequest = shouldBufferRequestBody(request)
+      ? await bufferRequestBody(request)
+      : request;
+
+    const staticResponse = await routeStatic(bufferedRequest, path, method);
+    if (staticResponse) {
+      return finish(
+        await applySecurityHeaders(staticResponse, isEmbeddablePath(path)),
+      );
+    }
+
+    seedEffectiveDomainHost(url);
+
+    const cleanLocation = trackingRedirectLocation(url, method);
+    if (cleanLocation) {
+      return finish(
+        new Response(null, {
+          headers: { location: cleanLocation },
+          status: 301,
+        }),
+      );
+    }
+
+    if (!isSetupPath(path)) settings.prefetchVersion();
+
+    const unavailable = await initializeDatabaseForPath(path);
+    if (unavailable) return finish(unavailable);
+
+    await prepareRequestEnvironment(url, path, method);
+
+    if (!isValidContentType(bufferedRequest, path)) {
+      return finish(contentTypeRejectionResponse());
+    }
+
+    response = finish(
+      await routeAndFinalize(bufferedRequest, url, path, method, server),
+    );
+    assertSettingsReadsDeclared(`${method} ${path}`);
+  } catch (error) {
+    response = finish(handleRoutingError(error, method, path));
+  } finally {
+    await flushPendingWork();
+  }
+  return response;
+};
