+++ src/features/app/routes.ts
@@ -1,0 +1,285 @@
+/* jscpd:ignore-start -- imports */
+import { once, reduce } from "#fp";
+import {
+  htmlResponse,
+  jsonResponse,
+  notFoundResponse,
+  redirectResponse,
+} from "#routes/response.ts";
+import {
+  createRouter,
+  defineRoutes,
+  type RouteHandlerFn,
+} from "#routes/router.ts";
+import { getPrefix } from "#routes/settings-bundles.ts";
+import type { PathMethodRoute, ServerContext } from "#routes/types.ts";
+import { settings } from "#shared/db/settings.ts";
+import { isReadOnly } from "#shared/env.ts";
+import type { ResponseHandler } from "#shared/response-steps.ts";
+import { readOnlyPage } from "#templates/public/errors.tsx";
+import { readOnlyBlock } from "./read-only.ts";
+
+/* jscpd:ignore-end */
+
+type RouterFn = ReturnType<typeof createRouter>;
+type CompletePathMethodRoute = (
+  ...args: Parameters<PathMethodRoute>
+) => Promise<Response>;
+type AppRouteRequest = {
+  method: string;
+  path: string;
+  request: Request;
+  server: ServerContext | undefined;
+};
+
+/** Give complete app routes one named request value instead of four arguments. */
+export const defineAppRoute =
+  (
+    handle: (route: AppRouteRequest) => Promise<Response>,
+  ): CompletePathMethodRoute =>
+  (request, path, method, server) =>
+    handle({ method, path, request, server });
+
+const READ_ONLY_MESSAGE = "This site is in read-only mode";
+
+const readOnlyResponses = {
+  api: (): Response => jsonResponse({ error: READ_ONLY_MESSAGE }, 403),
+  page: (): Response => redirectResponse("/read-only"),
+} as const;
+
+/** Lazily import a module once and pick one export from it. */
+const lazyExport = <M, K extends keyof M>(
+  load: () => Promise<M>,
+  key: K,
+): (() => Promise<M[K]>) => once(async () => (await load())[key]);
+
+/** Lazy-load a small route table assembled from named handlers. */
+const lazyRouter = <M>(
+  load: () => Promise<M>,
+  routes: (module: M) => Record<string, RouteHandlerFn>,
+): (() => Promise<RouterFn>) =>
+  once(async () => createRouter(routes(await load())));
+
+// Import specifiers stay literal so esbuild can bundle every lazy target.
+const routeLoaders = {
+  admin: lazyExport(() => import("#routes/admin/index.ts"), "routeAdmin"),
+  api: lazyExport(() => import("#routes/api/index.ts"), "routeApi"),
+  balance: lazyExport(
+    () => import("#routes/public/balance.ts"),
+    "routeBalance",
+  ),
+  checkin: lazyExport(() => import("#routes/checkin.ts"), "routeCheckin"),
+  demoReset: lazyExport(
+    () => import("#routes/admin/database-reset.ts"),
+    "routeDatabaseReset",
+  ),
+  feed: lazyExport(() => import("#routes/feeds.ts"), "routeFeed"),
+  googleWallet: lazyExport(
+    () => import("#routes/wallet/google.ts"),
+    "routeGoogleWallet",
+  ),
+  image: lazyExport(() => import("#routes/images.ts"), "routeImage"),
+  join: lazyExport(() => import("#routes/join.ts"), "routeJoin"),
+  news: lazyExport(() => import("#routes/public/news.ts"), "routeNews"),
+  order: lazyExport(() => import("#routes/public/order.ts"), "routeOrder"),
+  payment: lazyExport(() => import("#routes/api/webhooks.ts"), "routePayment"),
+  sitePage: lazyExport(
+    () => import("#routes/public/site-page.ts"),
+    "routeSitePage",
+  ),
+  smsWebhook: lazyExport(
+    () => import("#routes/api/sms-webhook.ts"),
+    "routeSmsWebhook",
+  ),
+  ticket: lazyExport(
+    () => import("#routes/public/ticket-routes.ts"),
+    "routeTicket",
+  ),
+  ticketView: lazyExport(
+    () => import("#routes/tickets/index.ts"),
+    "routeTicketView",
+  ),
+  wallet: lazyExport(() => import("#routes/wallet/index.ts"), "routeWallet"),
+  walletWebservice: lazyExport(
+    () => import("#routes/wallet/webservice.ts"),
+    "routeWalletWebservice",
+  ),
+};
+
+const handlerLoaders = {
+  customCss: lazyExport(
+    () => import("#routes/public/custom-css.ts"),
+    "handleCustomCss",
+  ),
+  orderJs: lazyExport(
+    () => import("#routes/public/order-js.ts"),
+    "handleOrderJs",
+  ),
+};
+
+const loadPublicPages = once(() => import("#routes/public/pages.ts"));
+
+const loadAttachmentRoutes = once(async () =>
+  createRouter((await import("#routes/attachments.ts")).attachmentRoutes),
+);
+
+const loadAdminApiRoutes = once(async () =>
+  createRouter((await import("#routes/admin/api.ts")).adminApiRoutes),
+);
+
+const loadScheduledRoutes = once(async () =>
+  createRouter((await import("#routes/scheduled.ts")).scheduledRoutes),
+);
+
+const loadInstanceRoutes = once(async () =>
+  createRouter((await import("#routes/instance.ts")).instanceRoutes),
+);
+
+const exactRouteLoaders = {
+  addressLookup: lazyRouter(
+    () => import("#routes/public/address-lookup.ts"),
+    ({ handleAddressLookupGet }) =>
+      defineRoutes({ "GET /address-lookup": handleAddressLookupGet }),
+  ),
+  renewal: lazyRouter(
+    () => import("#routes/public/renewal.ts"),
+    ({ handleRenewalGet, handleRenewalPost }) =>
+      defineRoutes({
+        "GET /renew": handleRenewalGet,
+        "POST /renew": handleRenewalPost,
+      }),
+  ),
+  unsubscribe: lazyRouter(
+    () => import("#routes/public/unsubscribe.ts"),
+    ({ handleUnsubscribeGet, handleUnsubscribePost }) =>
+      defineRoutes({
+        "GET /unsubscribe": handleUnsubscribeGet,
+        "POST /unsubscribe": handleUnsubscribePost,
+      }),
+  ),
+};
+
+const lazyRoute =
+  (load: () => Promise<RouterFn>): RouterFn =>
+  async (request, path, method, server) =>
+    (await load())(request, path, method, server);
+
+type PublicPagesModule = Awaited<ReturnType<typeof loadPublicPages>>;
+
+type PublicGetPageSpec = {
+  prefix: string;
+  pick: (pages: PublicPagesModule) => ResponseHandler<[request: Request]>;
+};
+
+const PUBLIC_GET_PAGES: PublicGetPageSpec[] = [
+  { pick: (pages) => pages.handleHome, prefix: "" },
+  { pick: (pages) => pages.handlePublicListings, prefix: "listings" },
+  { pick: (pages) => pages.handlePublicTerms, prefix: "terms" },
+];
+
+const publicPagePath = (prefix: string): string =>
+  prefix === "" ? "/" : `/${prefix}`;
+
+const publicPageHandlers = reduce(
+  (handlers: Record<string, RouterFn>, spec: PublicGetPageSpec) => {
+    const path = publicPagePath(spec.prefix);
+    handlers[spec.prefix] = async (request, requestPath, method) => {
+      if (requestPath !== path || method !== "GET") return null;
+      return spec.pick(await loadPublicPages())(request);
+    };
+    return handlers;
+  },
+  {},
+)(PUBLIC_GET_PAGES);
+
+const contactPrefixHandler: RouterFn = async (request, path, method) => {
+  if (path !== "/contact") return null;
+  const pages = await loadPublicPages();
+  if (method === "GET") return pages.handlePublicContact(request);
+  if (method === "POST") return pages.handlePublicContactSubmit(request);
+  return null;
+};
+
+const legacyEventsRedirectHandler: RouterFn = async (_request, path, method) =>
+  path === "/events" && method === "GET" && settings.features.site
+    ? redirectResponse("/listings")
+    : null;
+
+const orderJsPrefixHandler: RouterFn = async (request, path, method) => {
+  if (path !== "/order.js" || method !== "GET") return null;
+  return (await handlerLoaders.orderJs())(request);
+};
+
+const customCssPrefixHandler: RouterFn = async (_request, path, method) => {
+  if (path !== "/custom.css" || method !== "GET") return null;
+  return (await handlerLoaders.customCss())();
+};
+
+const prefixHandlers: Record<string, RouterFn> = {
+  ...publicPageHandlers,
+  "address-lookup": lazyRoute(exactRouteLoaders.addressLookup),
+  admin: lazyRoute(routeLoaders.admin),
+  api: async (request, path, method, server) => {
+    const adminResult = await (await loadAdminApiRoutes())(
+      request,
+      path,
+      method,
+      server,
+    );
+    if (adminResult) return adminResult;
+    return settings.showPublicApi
+      ? (await routeLoaders.api())(request, path, method, server)
+      : null;
+  },
+  attachment: lazyRoute(loadAttachmentRoutes),
+  calculate: lazyRoute(routeLoaders.ticket),
+  caldav: lazyRoute(routeLoaders.feed),
+  checkin: lazyRoute(routeLoaders.checkin),
+  contact: contactPrefixHandler,
+  "custom.css": customCssPrefixHandler,
+  demo: lazyRoute(routeLoaders.demoReset),
+  events: legacyEventsRedirectHandler,
+  feeds: lazyRoute(routeLoaders.feed),
+  gwallet: lazyRoute(routeLoaders.googleWallet),
+  image: lazyRoute(routeLoaders.image),
+  instance: lazyRoute(loadInstanceRoutes),
+  join: lazyRoute(routeLoaders.join),
+  news: lazyRoute(routeLoaders.news),
+  order: lazyRoute(routeLoaders.order),
+  "order.js": orderJsPrefixHandler,
+  page: lazyRoute(routeLoaders.sitePage),
+  pay: lazyRoute(routeLoaders.balance),
+  payment: lazyRoute(routeLoaders.payment),
+  "read-only": (_request, path, method) =>
+    path === "/read-only" && method === "GET"
+      ? Promise.resolve(htmlResponse(readOnlyPage()))
+      : Promise.resolve(null),
+  renew: lazyRoute(exactRouteLoaders.renewal),
+  scheduled: lazyRoute(loadScheduledRoutes),
+  sms: lazyRoute(routeLoaders.smsWebhook),
+  t: lazyRoute(routeLoaders.ticketView),
+  ticket: lazyRoute(routeLoaders.ticket),
+  unsubscribe: lazyRoute(exactRouteLoaders.unsubscribe),
+  v1: lazyRoute(routeLoaders.walletWebservice),
+  wallet: lazyRoute(routeLoaders.wallet),
+};
+
+/** Route a request after database and settings setup have completed. */
+export const routeMainApp = async ({
+  request,
+  path,
+  method,
+  server,
+}: AppRouteRequest): Promise<Response> => {
+  if (isReadOnly()) {
+    const blocked = readOnlyBlock(path, method);
+    if (blocked) return readOnlyResponses[blocked]();
+  }
+
+  const prefix = getPrefix(path);
+  if (!Object.hasOwn(prefixHandlers, prefix)) return notFoundResponse();
+  return (
+    (await prefixHandlers[prefix]?.(request, path, method, server)) ??
+    notFoundResponse()
+  );
+};
